refactor frontend scripts
